requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
black==19.10b0
bandit==1.7.0
click==7.1.2
colorama==0.4.4
coverage==5.5
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
importlib-metadata==3.7.3
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytz==2021.1
regex==2021.3.17
six==1.15.0
toml==0.10.2
tox==3.23.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
virtualenv==20.4.3
Werkzeug==1.0.1
zipp==3.4.1
simple-orm==0.0.1
